Routines not available via Pkg

MParT.Serialize — Function

Serialize(obj, filename)

Serializes obj into file filename. REQUIRES CEREAL INSTALLATION.

MParT.Deserialize — Function

Deserialize(obj, filename)

Deserializes filename and puts the contents in obj. REQUIRES CEREAL INSTALLATION.

The object obj can be of type MapOptions or FixedMultiIndexSet. This will create a new pointer– other objects with the same pointer will not be modified, but the contents of obj will now point to the deserialized object.

MParT.DeserializeMap — Function

DeserializeMap(filename::String)

REQUIRES CEREAL INSTALLATION. Deserializes a map and returns its input dimension, output dimension, and coefficient.