Basic map Routines

MParT.Evaluate — Function

Evaluate(map, points)

Evaluates the function map at points, where each column of points is a different sample.

If map $:\mathbb{R}^m\to\mathbb{R}^n$, then points $\in\mathbb{R}^m\times\mathbb{R}^k$, where $k$ is the number of points.

MParT.Inverse — Function

Inverse(map, y, x)

If map represents function $T(y,x)$, then this function calculates $T(y,\cdot)^{-1}(x)$.

If map is square, you still require y, but it can be a 0 x k matrix.

MParT.ComposedMap — Type

ComposedMap(maps::Vector)

Creates a ComposedMap from a vector of ConditionalMapBase objects.

MParT.MapOptions — Function

MapOptions(;kwargs...)

Creates options for parameterized map.

All possible keyword arguments are in example, with some important arguments described below. See C++ documentation for an exhaustive description.

Arguments

basisType::String: Includes "ProbabilistHermite", "PhysicistHermite", "HermiteFunctions"

basisLB::Float64,basisUB::Float64: The bounds for where we start linearizing the map. These default to infinities, but often making the data near the origin and setting them to a small finite number (e.g. +-3) works well.

Example

julia> MapOptions(basisType="HermiteFunctions", basisLB=-3., basisUB=3.)
basisType = HermiteFunctions
basisLB = -3
basisUB = 3
basisNorm = true
posFuncType = SoftPlus
quadType = AdaptiveSimpson
quadAbsTol = 1e-06
quadRelTol = 1e-06
quadMaxSub = 30
quadMinSub = 0
quadPts = 5
contDeriv = true

See also CreateComponent, TriangularMap, CreateTriangular

MParT.TriangularMap — Type

TriangularMap(maps::Vector)

Creates a TriangularMap from a vector of ConditionalMapBase objects.

Examples

julia> dim, order = 5, 3;

julia> msets = [FixedMultiIndexSet(d, order) for d in 1:dim];

julia> opts = MapOptions();

julia> components = [CreateComponent(mset, opts) for mset in msets];

julia> trimap = TriangularMap(components);

MParT.CreateTriangular — Function

CreateTriangular(inDim::Int, outDim::Int, p::Int, opts::MapOptions)

Creates a total order p map with dimensions inDim and outDim with specifications opts.

MParT.CreateComponent — Function

CreateComponent(mset::FixedMultiIndexSet, opts::MapOptions)

Create a single-output component with approximation order given by mset and specified by opts