MParT

Documentation for MParT.jl. This documentation is incomplete and only features some basic functionality. For documentation of all functions as well as source installation instructions, other binding information, and many examples, please look at the C++ documentation. If there's a function there unavailable in Julia, please let the developers know via a GitHub issue.

MultiIndex-related routines

Basic map Routines

Map Training Routines

Adaptive Map Training Routines

Routines not available via Pkg

General Utilities

MParT.Concurrency — Function

Concurrency()

See how many threads MParT is using.