MParT

Documentation for MParT.

MParT.ComposedMap

MParT.MultiIndex

MParT.MultiIndexSet

MParT.TriangularMap

MParT.ATMOptions

MParT.CreateGaussianKLObjective

MParT.DeserializeMap

MParT.MapOptions

MParT.TrainMapAdaptive

MParT.TrainOptions

MParT.ComposedMap — Method

`ComposedMap(maps::Vector)`

Creates a ComposedMap from a vector of ConditionalMapBase objects.

MParT.MultiIndex — Method

MultiIndex(A::AbstractVector{<:Int})

MultiIndex defines the order of one term of a function expansion for each dimensions.

Example

julia> degrees = [3,2,1]; # Represents polynomial basis function similar to x^3y^2z^1

julia> midx = MultiIndex(degrees);

julia> midx[3]
0x00000001

See also MultiIndexSet, FixedMultiIndexSet, Fix

MParT.MultiIndexSet — Method

MultiIndexSet(A::AbstractVecOrMat{<:Integer})

Create a set of MultiIndices from the rows of A.

These indices represent a basis for a multivariate expansion or, further, monotone expansion. Each element of the set is a MultiIndex representing one basis function via the degrees in each dimension.

Example

julia> # Functions like: c_1xy^2z^3 + c_2xyz + c_3

julia> A = [1 2 3;1 1 1; 0 0 0];

julia> mset = MultiIndexSet(A);

See alsoMultiIndex, FixedMultiIndex, Fix

MParT.TriangularMap — Method

`TriangularMap(maps::Vector)`

Creates a TriangularMap from a vector of ConditionalMapBase objects.

Examples

julia> dim, order = 5, 3;

julia> msets = [FixedMultiIndexSet(d, order) for d in 1:dim];

julia> opts = MapOptions();

julia> components = [CreateComponent(mset, opts) for mset in msets];

julia> trimap = TriangularMap(components);

MParT.ATMOptions — Method

`ATMOptions(;kwargs...)`

Takes the fields from MParT's ATMOptions as keyword arguments, and assigns the field value based on a String from keyword arguments.

Examples

julia> maxDegrees = MultiIndex(2,3); # limit both dimensions by order 3

julia> ATMOptions(opt_alg="LD_SLSQP", maxDegrees=maxDegrees);

MParT.CreateGaussianKLObjective — Function

CreateGaussianKLObjective(train)
CreateGaussianKLObjective(train, outputDim)
CreateGaussianKLObjective(train, test)
CreateGaussianKLObjective(train, test, outputDim)

Create an objective for the optimization problem of training a transport map.

Currently only supports variational simulation-based inference, i.e. creating a map that uses samples train from target distribution and a Gaussian reference distribution of dimension outputDim. If outputDim is zero, this is equivalent to when outputDim == size(train,1).

Arguments

train::Matrix{Float64}: mandatory training dataset

test::Matrix{Float64}: optional test dataset

outputDim::Int = 0: Dimensions of output of map

Examples

julia> using Random; rng = Random.Xoshiro(1);

julia> # Replace RNG type with MersenneTwister for <1.7

julia> N, inDim, outDim = 1000, 4, 2;

julia> samples = 2*randn(rng, inDim, N) .+ 5;

julia> obj1 = CreateGaussianKLObjective(samples);

julia> obj2 = CreateGaussianKLObjective(samples, outDim);

julia> train, test = samples[:,1:500], samples[:,501:end];

julia> obj3 = CreateGaussianKLObjective(train, test);

julia> obj4 = CreateGaussianKLObjective(train, test, outDim);

MParT.DeserializeMap — Method

DeserializeMap(filename::String)

REQUIRES CEREAL INSTALLATION. Deserializes a map and returns its input dimension, output dimension, and coefficient.

MParT.MapOptions — Method

MapOptions(;kwargs...)

Takes the fields from MParT's MapOptions as keyword arguments, and assigns the field value based on a String from the kwarg value.

Arguments

``

Example

julia> MapOptions(basisType="HermiteFunctions", basisLB=-3., basisUB=3.)
basisType = HermiteFunctions
basisLB = -3
basisUB = 3
basisNorm = true
posFuncType = SoftPlus
quadType = AdaptiveSimpson
quadAbsTol = 1e-06
quadRelTol = 1e-06
quadMaxSub = 30
quadMinSub = 0
quadPts = 5
contDeriv = true

MParT.TrainMapAdaptive — Method

TrainMapAdaptive(msets, objective, options)

This implements the ATM algorithm (Baptista et al. 2022), which takes in an initial guess of multiindex sets for each output dimension and adapts those sets to better suit the probability distribution.

Examples

MParT.TrainOptions — Method

`TrainOptions(;kwargs...)`

Takes the fields from MParT's TrainOptions as keyword arguments, and assigns the field value based on a String from keyword arguments.

Arguments

-``:

Examples

julia> TrainOptions(opt_alg="LD_SLSQP", opt_maxeval = 1_000_000)
opt_alg = LD_SLSQP
opt_stopval = -inf
opt_ftol_rel = 0.001
opt_ftol_abs = 0.001
opt_xtol_rel = 0.0001
opt_xtol_abs = 0.0001
opt_maxeval = 1000000
opt_maxtime = inf
verbose = 0